+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.50] - 2025-01-07
 ### Added
 - BOS options: bss_read_timeout, hsm_read_timeout, capmc_read_timeout.
   Allow the amount of time BOS waits for a response from these services before